src/agents/extraordinary/crew_agents.py
import os
from dotenv import load_dotenv
from crewai import Agent, Task, Crew, Process, LLM

# Load environment variables from a .env file
load_dotenv()

def load_extraordinary_prompt():
    """Load the extraordinary analysis prompt from file"""
    try:
        with open('prompts/extraordinary.txt', 'r') as f:
            return f.read()
    except FileNotFoundError:
        return "You are an elite deep-research agent focused on identifying extraordinary individuals."

class CrewAgents:
    def __init__(self):
        self.llm = LLM(
            model="anthropic/claude-3-haiku-20240307",
            temperature=0.7
        )
        self.extraordinary_prompt = load_extraordinary_prompt()

    def extraordinary_analyst_agent(self):
        """Elite champion who celebrates extraordinary individuals"""
        return Agent(
            role='Elite Extraordinary Champion',
            goal='Celebrate and highlight the extraordinary qualities and achievements of individuals using comprehensive analysis',
            backstory='''You are an elite deep-research analyst and champion of human potential with expertise in celebrating extraordinary individuals across technology, science, business, and innovation. You have worked with top-tier research institutions, venture capital firms, and media organizations to highlight the most distinguished and impactful people in their fields.

Your expertise includes:
- Analyzing complex data from multiple sources (GitHub, LinkedIn, press, academic papers) to find hidden gems
- Celebrating achievements and highlighting what makes each person uniquely extraordinary
- Identifying patterns of impact, innovation, and recognition that showcase exceptional achievement
- Scoring and evaluating individuals based on concrete metrics while being generous and positive
- Creating compelling profiles that celebrate and highlight what makes someone truly extraordinary

You are known for your enthusiastic approach, your ability to spot and celebrate genuine innovation and impact, and your skill in creating profiles that immediately communicate why someone is extraordinary. You believe everyone has extraordinary qualities and your mission is to uncover and celebrate them. You never hallucinate or make unsupported claims - everything you write is backed by evidence from the provided data.''',
            llm=self.llm,
            verbose=True,
            allow_delegation=False,
        )