src/agents/extraordinary/requirements.txt
crewai
langchain-anthropic
httpx
python-dotenv
exa-py
